Refine architecture boundaries
