+++ server/repositories/revisions.ts
@@ -1,7 +1,7 @@
 import { existsSync } from "node:fs";
 import { readdir } from "node:fs/promises";
 import path from "node:path";
-import type { PolicyRevision, RevisionProposal, WorkspaceRevision } from "../../src/types";
+import type { PolicyRevision, RevisionProposal, WorkspaceRevision } from "../../shared/types";
 import { readJson, writeJson } from "../util";
 
 export interface RevisionRepository {
