+++ server/repositories/sweeps.ts
@@ -1,7 +1,7 @@
 import { existsSync } from "node:fs";
 import { readdir } from "node:fs/promises";
 import path from "node:path";
-import type { SweepBatch, SweepFeedbackTrace, SweepState } from "../../src/types";
+import type { SweepBatch, SweepFeedbackTrace, SweepState } from "../../shared/types";
 import { readJson, writeJson } from "../util";
 
 export interface SweepRepository {
