+++ tsconfig.json
@@ -11,5 +11,5 @@
     "noEmit": true,
     "types": ["node", "bun"]
   },
-  "include": ["src", "server", "scripts", "server.ts", "cli.ts", "attention.ts", "vite.config.ts"]
+  "include": ["src", "server", "scripts", "shared", "server.ts", "cli.ts", "attention.ts", "vite.config.ts"]
 }
